--- src/datadoc/backend/tests/utility/__init__.py
@@ -1,1 +1,0 @@
-"""Utility files for tests."""
